backend/pip-requirements.txt
pandas==2.2.3
tqdm==4.67.1
weaviate_client==4.10.4
flask
flask-cors